refactor device in sycl_device, restore ctx in create_queue
